Move API elements up one level
