+++ lib/api.js
@@ -56,7 +56,7 @@ var API = module.exports = function API(opts) {
   });
 
   // load route definitions
-  this.express.use("/api", require("./api/routes"));
+  this.express.use("/api", require("./routes"));
 
   // error handling
   this.express.use(function(err, req, res, next) {
@@ -71,8 +71,8 @@ API.prototype.defaults = {
   auth: false,
   CORS: "",
   ssl: {
-    key: path.normalize(__dirname + "/api/ssl/server.key"),
-    cert: path.normalize(__dirname + "/api/ssl/server.crt")
+    key: path.normalize(__dirname + "/ssl/server.key"),
+    cert: path.normalize(__dirname + "/ssl/server.crt")
   }
 };
 
@@ -101,7 +101,7 @@ API.prototype.setupAuth = function setupAuth() {
 
   if (!!this.auth && typeof(this.auth) === "object" && this.auth.type) {
     var type = this.auth.type,
-        module = "./api/auth/" + type,
+        module = "./auth/" + type,
         filename = path.normalize(__dirname + "/" + module + ".js"),
         exists = fs.existsSync(filename);
