+++ lib/auth/basic.js
